Represent a player's balance using a float

+++ src/player.ml
@@ -1,6 +1,6 @@
 type t = {
     name: string;
-    funds: int;
+    funds: float;
   }
 
 let create name funds = { name = name; funds = funds }
@@ -8,4 +8,4 @@ let create name funds = { name = name; funds = funds }
 let serialize (player : t) =
   let open Yojson.Basic in
   `Assoc [("name", `String player.name);
-         ("funds", `Int player.funds)]
+         ("funds", `Float player.funds)]
